requirements.txt
dash==2.0.0
dash-bootstrap-components==1.0.0
dash-core-components==2.0.0
dash-daq==0.5.0
dash-html-components==2.0.0
dash-table==5.0.0
numpy==1.20.1
pandas==1.2.4
mysql-connector-python==8.0.27